Added definitions for CodeParser class
The class and Makefile are in place in the codeparser directory. A few
methods that I think will be needed have been defined and currently have
placeholder return values to stop warnings when I was testing the build.
This is not linked at all anywhere else in the code, including the main
build system! Don't do that until it is implemented and ready to be
used.