Renamed top level build system files
Having to specify the name of the Makefile that you used the most got
kind of cumbersome, so to make things more intuitive, the files at the
top level have been renamed. Now the file to do a full object build is
named Makefile, and the file to do the linking is named linker.mk.