Tentative new UI setup
I've arranged the UI roughly how it needs to look. There might be a
couple little proportion issues that need to be worked out, but the
general look is there. The input/output boxes are currently acting like
the AR/TST output boxes used to; this is because I was only rearranging
the UI and wanted to keep the controller in place and not disturb it.

refs issue #3