FileHandler automatically loads previous location
If the user has selected a file already during this run of the program,
the second time that they open the file dialog it will be at the
location of the previous file, rather than making them navigate from the
home directory again. The default is still the home directory.