Removed excessive newlines from the end of files
It turns out that Vim auto-EOFs files, and I had no idea about this when
I started the project. Ever since I found out I've wanted to fix this,
so here it is. I also modified the page_generator not to output an extra
newline at the end of the files generated.